Move initial tests out of AbstractTestQueries
