+++ presto-tests/src/main/java/com/facebook/presto/tests/AbstractTestQueries.java
@@ -2082,69 +2082,6 @@ public abstract class AbstractTestQueries
     }
 
     @Test
-    public void testExcept()
-    {
-        assertQuery(
-                "SELECT regionkey FROM nation WHERE nationkey < 7 " +
-                        "EXCEPT select regionkey FROM nation WHERE nationkey > 21");
-        assertQuery(
-                "SELECT regionkey FROM nation WHERE nationkey < 7 " +
-                        "EXCEPT DISTINCT SELECT regionkey FROM nation WHERE nationkey > 21",
-                "VALUES 0, 4");
-        assertQuery(
-                "WITH wnation AS (SELECT nationkey, regionkey FROM nation) " +
-                        "SELECT regionkey FROM wnation WHERE nationkey < 7 " +
-                        "EXCEPT SELECT regionkey FROM wnation WHERE nationkey > 21",
-                "VALUES 0, 4");
-        assertQuery(
-                "SELECT num FROM (SELECT 1 as num FROM nation WHERE nationkey=10 " +
-                        "EXCEPT SELECT 2 FROM nation WHERE nationkey=20) T");
-        assertQuery(
-                "SELECT nationkey, nationkey / 2 FROM (SELECT nationkey FROM nation WHERE nationkey < 10 " +
-                        "EXCEPT SELECT nationkey FROM nation WHERE nationkey > 4) T WHERE nationkey % 2 = 0");
-        assertQuery(
-                "SELECT regionkey FROM (SELECT regionkey FROM nation WHERE nationkey < 7 " +
-                        "EXCEPT SELECT regionkey FROM nation WHERE nationkey > 21) " +
-                        "UNION SELECT 3");
-        assertQuery(
-                "SELECT regionkey FROM (SELECT regionkey FROM nation WHERE nationkey < 7 " +
-                        "UNION SELECT regionkey FROM nation WHERE nationkey > 21) " +
-                        "EXCEPT SELECT 1");
-        assertQuery(
-                "SELECT regionkey FROM (SELECT regionkey FROM nation WHERE nationkey < 7 " +
-                        "EXCEPT SELECT regionkey FROM nation WHERE nationkey > 21) " +
-                        "UNION ALL SELECT 4");
-        assertQuery(
-                "SELECT * FROM (VALUES 1, 2) " +
-                        "EXCEPT SELECT * FROM (VALUES 3.0, 2)");
-        assertQuery("SELECT NULL, NULL EXCEPT SELECT NULL, NULL FROM nation");
-
-        assertQuery(
-                "(SELECT * FROM (VALUES 1) EXCEPT SELECT * FROM (VALUES 0))" +
-                        "EXCEPT (SELECT * FROM (VALUES 1) EXCEPT SELECT * FROM (VALUES 1))");
-
-        MaterializedResult emptyResult = computeActual("SELECT 0 EXCEPT (SELECT regionkey FROM nation WHERE nationkey <10)");
-        assertEquals(emptyResult.getMaterializedRows().size(), 0);
-    }
-
-    @Test
-    public void testExceptWithAggregation()
-    {
-        assertQuery("SELECT COUNT(*) FROM nation EXCEPT SELECT COUNT(regionkey) FROM nation where regionkey < 3 HAVING SUM(regionkey) IS NOT NULL");
-        assertQuery("SELECT SUM(nationkey), COUNT(name) FROM (SELECT nationkey, name FROM nation where nationkey < 6 EXCEPT SELECT regionkey, name FROM nation) n");
-        assertQuery("(SELECT SUM(nationkey) FROM nation GROUP BY regionkey ORDER BY 1 LIMIT 2) EXCEPT SELECT COUNT(*) * 2 FROM nation");
-        assertQuery("SELECT COUNT(a) FROM (SELECT nationkey AS a FROM (SELECT nationkey FROM nation EXCEPT SELECT regionkey FROM nation) n1 EXCEPT SELECT regionkey FROM nation) n2");
-        assertQuery("SELECT COUNT(*), SUM(2), regionkey FROM (SELECT nationkey, regionkey FROM nation EXCEPT SELECT regionkey, regionkey FROM nation) n GROUP BY regionkey HAVING regionkey < 3");
-        assertQuery("SELECT COUNT(*) FROM (SELECT nationkey FROM nation EXCEPT SELECT 10) n1 EXCEPT SELECT regionkey FROM nation");
-    }
-
-    @Test
-    public void testExceptAllFails()
-    {
-        assertQueryFails("SELECT * FROM (VALUES 1, 2, 3, 4) EXCEPT ALL SELECT * FROM (VALUES 3, 4)", "line 1:35: EXCEPT ALL not yet implemented");
-    }
-
-    @Test
     public void testCountAllWithComparison()
     {
         assertQuery("SELECT COUNT(*) FROM lineitem WHERE tax < discount");
@@ -4388,226 +4325,6 @@ public abstract class AbstractTestQueries
                 .build();
 
         assertEquals(actual, expected);
-    }
-
-    @Test
-    public void testRowNumberNoOptimization()
-    {
-        MaterializedResult actual = computeActual("" +
-                "SELECT orderkey, orderstatus FROM (\n" +
-                "   SELECT row_number() OVER () rn, orderkey, orderstatus\n" +
-                "   FROM orders\n" +
-                ") WHERE NOT rn <= 10");
-        MaterializedResult all = computeExpected("SELECT orderkey, orderstatus FROM ORDERS", actual.getTypes());
-        assertEquals(actual.getMaterializedRows().size(), all.getMaterializedRows().size() - 10);
-        assertContains(all, actual);
-
-        actual = computeActual("" +
-                "SELECT orderkey, orderstatus FROM (\n" +
-                "   SELECT row_number() OVER () rn, orderkey, orderstatus\n" +
-                "   FROM orders\n" +
-                ") WHERE rn - 5 <= 10");
-        all = computeExpected("SELECT orderkey, orderstatus FROM ORDERS", actual.getTypes());
-        assertEquals(actual.getMaterializedRows().size(), 15);
-        assertContains(all, actual);
-    }
-
-    @Test
-    public void testRowNumberLimit()
-    {
-        MaterializedResult actual = computeActual("" +
-                "SELECT row_number() OVER (PARTITION BY orderstatus) rn, orderstatus\n" +
-                "FROM orders\n" +
-                "LIMIT 10");
-        assertEquals(actual.getMaterializedRows().size(), 10);
-
-        actual = computeActual("" +
-                "SELECT row_number() OVER (PARTITION BY orderstatus ORDER BY orderkey) rn\n" +
-                "FROM orders\n" +
-                "LIMIT 10");
-        assertEquals(actual.getMaterializedRows().size(), 10);
-
-        actual = computeActual("" +
-                "SELECT row_number() OVER () rn, orderstatus\n" +
-                "FROM orders\n" +
-                "LIMIT 10");
-        assertEquals(actual.getMaterializedRows().size(), 10);
-
-        actual = computeActual("" +
-                "SELECT row_number() OVER (ORDER BY orderkey) rn\n" +
-                "FROM orders\n" +
-                "LIMIT 10");
-        assertEquals(actual.getMaterializedRows().size(), 10);
-    }
-
-    @Test
-    public void testRowNumberMultipleFilters()
-    {
-        MaterializedResult actual = computeActual("" +
-                "SELECT * FROM (" +
-                "   SELECT a, row_number() OVER (PARTITION BY a ORDER BY a) rn\n" +
-                "   FROM (VALUES (1), (1), (1), (2), (2), (3)) t (a)) t " +
-                "WHERE rn < 3 AND rn % 2 = 0 AND a = 2 LIMIT 2");
-        MaterializedResult expected = resultBuilder(getSession(), BIGINT, BIGINT)
-                .row(2, 2L)
-                .build();
-        assertEqualsIgnoreOrder(actual.getMaterializedRows(), expected.getMaterializedRows());
-    }
-
-    @Test
-    public void testRowNumberFilterAndLimit()
-    {
-        MaterializedResult actual = computeActual("" +
-                "SELECT * FROM (" +
-                "SELECT a, row_number() OVER (PARTITION BY a ORDER BY a) rn\n" +
-                "FROM (VALUES (1), (2), (1), (2)) t (a)) t WHERE rn < 2 LIMIT 2");
-
-        MaterializedResult expected = resultBuilder(getSession(), BIGINT, BIGINT)
-                .row(1, 1L)
-                .row(2, 1L)
-                .build();
-        assertEqualsIgnoreOrder(actual.getMaterializedRows(), expected.getMaterializedRows());
-
-        actual = computeActual("" +
-                "SELECT * FROM (" +
-                "SELECT a, row_number() OVER (PARTITION BY a) rn\n" +
-                "FROM (VALUES (1), (2), (1), (2), (1)) t (a)) t WHERE rn < 3 LIMIT 2");
-
-        expected = resultBuilder(getSession(), BIGINT, BIGINT)
-                .row(1, 1L)
-                .row(1, 2L)
-                .row(2, 1L)
-                .row(2, 2L)
-                .build();
-        assertEquals(actual.getMaterializedRows().size(), 2);
-        assertContains(expected, actual);
-    }
-
-    @Test
-    public void testRowNumberUnpartitionedFilter()
-    {
-        MaterializedResult actual = computeActual("" +
-                "SELECT orderkey, orderstatus FROM (\n" +
-                "   SELECT row_number() OVER () rn, orderkey, orderstatus\n" +
-                "   FROM orders\n" +
-                ") WHERE rn <= 5 AND orderstatus != 'Z'");
-        MaterializedResult all = computeExpected("SELECT orderkey, orderstatus FROM ORDERS", actual.getTypes());
-        assertEquals(actual.getMaterializedRows().size(), 5);
-        assertContains(all, actual);
-
-        actual = computeActual("" +
-                "SELECT orderkey, orderstatus FROM (\n" +
-                "   SELECT row_number() OVER () rn, orderkey, orderstatus\n" +
-                "   FROM orders\n" +
-                ") WHERE rn < 5");
-        all = computeExpected("SELECT orderkey, orderstatus FROM ORDERS", actual.getTypes());
-
-        assertEquals(actual.getMaterializedRows().size(), 4);
-        assertContains(all, actual);
-
-        actual = computeActual("" +
-                "SELECT orderkey, orderstatus FROM (\n" +
-                "   SELECT row_number() OVER () rn, orderkey, orderstatus\n" +
-                "   FROM orders\n" +
-                ") LIMIT 5");
-        all = computeExpected("SELECT orderkey, orderstatus FROM ORDERS", actual.getTypes());
-
-        assertEquals(actual.getMaterializedRows().size(), 5);
-        assertContains(all, actual);
-    }
-
-    @Test
-    public void testRowNumberPartitionedFilter()
-    {
-        MaterializedResult actual = computeActual("" +
-                "SELECT orderkey, orderstatus FROM (\n" +
-                "   SELECT row_number() OVER (PARTITION BY orderstatus) rn, orderkey, orderstatus\n" +
-                "   FROM orders\n" +
-                ") WHERE rn <= 5");
-        MaterializedResult all = computeExpected("SELECT orderkey, orderstatus FROM ORDERS", actual.getTypes());
-
-        // there are 3 distinct orderstatus, so expect 15 rows.
-        assertEquals(actual.getMaterializedRows().size(), 15);
-        assertContains(all, actual);
-
-        // Test for unreferenced outputs
-        actual = computeActual("" +
-                "SELECT orderkey FROM (\n" +
-                "   SELECT row_number() OVER (PARTITION BY orderstatus) rn, orderkey\n" +
-                "   FROM orders\n" +
-                ") WHERE rn <= 5");
-        all = computeExpected("SELECT orderkey FROM ORDERS", actual.getTypes());
-
-        // there are 3 distinct orderstatus, so expect 15 rows.
-        assertEquals(actual.getMaterializedRows().size(), 15);
-        assertContains(all, actual);
-    }
-
-    @Test
-    public void testRowNumberJoin()
-    {
-        MaterializedResult actual = computeActual("SELECT a, rn\n" +
-                "FROM (\n" +
-                "    SELECT a, row_number() OVER (ORDER BY a) rn\n" +
-                "    FROM (VALUES (1), (2)) t (a)\n" +
-                ") a\n" +
-                "JOIN (VALUES (2)) b (b) ON a.a = b.b\n" +
-                "LIMIT 1");
-
-        MaterializedResult expected = resultBuilder(getSession(), BIGINT, BIGINT)
-                .row(2, 2L)
-                .build();
-        assertEqualsIgnoreOrder(actual.getMaterializedRows(), expected.getMaterializedRows());
-
-        actual = computeActual("SELECT a, rn\n" +
-                "FROM (\n" +
-                "    SELECT a, row_number() OVER (PARTITION BY a ORDER BY a) rn\n" +
-                "    FROM (VALUES (1), (2), (1), (2)) t (a)\n" +
-                ") a\n" +
-                "JOIN (VALUES (2)) b (b) ON a.a = b.b\n" +
-                "LIMIT 2");
-
-        expected = resultBuilder(getSession(), BIGINT, BIGINT)
-                .row(2, 1L)
-                .row(2, 2L)
-                .build();
-        assertEqualsIgnoreOrder(actual.getMaterializedRows(), expected.getMaterializedRows());
-    }
-
-    @Test
-    public void testRowNumberUnpartitionedFilterLimit()
-    {
-        assertQuery("" +
-                "SELECT row_number() OVER ()\n" +
-                "FROM lineitem JOIN orders ON lineitem.orderkey = orders.orderkey\n" +
-                "WHERE orders.orderkey = 10000\n" +
-                "LIMIT 20");
-    }
-
-    @Test
-    public void testRowNumberPropertyDerivation()
-    {
-        MaterializedResult actual = computeActual("" +
-                "SELECT orderkey, orderstatus, SUM(rn) OVER (PARTITION BY orderstatus) c\n" +
-                "FROM (\n" +
-                "   SELECT orderkey, orderstatus, row_number() OVER (PARTITION BY orderstatus) rn\n" +
-                "   FROM (\n" +
-                "       SELECT * FROM orders ORDER BY orderkey LIMIT 10\n" +
-                "   )\n" +
-                ")");
-        MaterializedResult expected = resultBuilder(getSession(), BIGINT, VARCHAR, BIGINT)
-                .row(1L, "O", 21L)
-                .row(2L, "O", 21L)
-                .row(3L, "F", 10L)
-                .row(4L, "O", 21L)
-                .row(5L, "F", 10L)
-                .row(6L, "F", 10L)
-                .row(7L, "O", 21L)
-                .row(32L, "O", 21L)
-                .row(33L, "F", 10L)
-                .row(34L, "O", 21L)
-                .build();
-        assertEqualsIgnoreOrder(actual.getMaterializedRows(), expected.getMaterializedRows());
     }
 
     @Test
